Split test constants into a separate file.
